requirements.txt
mne==0.23.4
numpy==1.21.2
pandas==1.3.3
patool==1.12
python-dateutil==2.8.2
pytz==2021.1
scipy==1.7.1
six==1.16.0
